:bug: カーソルがリンクの左隣にいると入力補完が開始されてしまっていた
